优化解决防范

+++ XXBRefreshDemo/XXBRefreshDemo/ViewController.m
@@ -75,7 +75,7 @@
     dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(1.0 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
         NSInteger count = self.dataSouceArray.count;
         
-        for (int i = count; i < count + 5; i++) {
+        for (NSInteger i = count; i < count + 5; i++) {
             [_dataSouceArray addObject:[NSString stringWithFormat:@"cell >>> %@",@(_dataSouceArray.count)]];
             [self.tableView insertRowsAtIndexPaths:@[[NSIndexPath indexPathForRow:count inSection:0]] withRowAnimation:UITableViewRowAnimationBottom];
         }
